Show message on ogcserver delete if contains layers

+++ admin/c2cgeoportal_admin/views/ogc_servers.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (c) 2017-2020, Camptocamp SA
+# Copyright (c) 2017-2021, Camptocamp SA
 # All rights reserved.
 
 # Redistribution and use in source and binary forms, with or without
@@ -31,7 +31,7 @@
 from functools import partial
 
 from c2cgeoform.schema import GeoFormSchemaNode
-from c2cgeoform.views.abstract_views import AbstractViews, ItemAction, ListField
+from c2cgeoform.views.abstract_views import AbstractViews, ItemAction, ListField, UserMessage
 from deform.widget import FormWidget
 from pyramid.httpexceptions import HTTPNotFound
 from pyramid.view import view_config, view_defaults
@@ -64,6 +64,14 @@ class OGCServerViews(AbstractViews):
     _id_field = "id"
     _model = OGCServer
     _base_schema = base_schema
+
+    MSG_COL = {
+        **AbstractViews.MSG_COL,
+        "cannot_delete": UserMessage(
+            _("Impossible to delete this server while it contains WMS layers."),
+            "alert-danger",
+        ),
+    }
 
     @view_config(route_name="c2cgeoform_index", renderer="../templates/index.jinja2")
     def index(self):
@@ -108,6 +116,17 @@ class OGCServerViews(AbstractViews):
 
     @view_config(route_name="c2cgeoform_item", request_method="DELETE", renderer="fast_json")
     def delete(self):
+        obj = self._get_object()
+        if len(obj.layers) > 0:
+            return {
+                "success": True,
+                "redirect": self._request.route_url(
+                    "c2cgeoform_item",
+                    action="edit",
+                    id=obj.id,
+                    _query=[("msg_col", "cannot_delete")],
+                ),
+            }
         return super().delete()
 
     @view_config(
